If revision does not exist, return latest attachment
If a request is made to the `GET /{db}/{doc}/{attachment}?rev={revid}`
for a revision that the remote database has no knowledge off, the latest
attachment should be returned to the requesting client.  `404 missing`
should not be returned unless the document itself is missing (i.e. not
the document revision is missing)